main_module/requirements.txt
absl-py==0.13.0
asgiref==3.4.1
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
Django==3.2.5
Flask==2.0.1
future==0.18.2
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.39.0
h5py==2.10.0
idna==2.10
imutils==0.5.3
iso8601==0.1.16
itsdangerous==2.0.1
Jinja2==3.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
livereload==2.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.2.1
numpy==1.18.2
oauthlib==3.1.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
qrcode==7.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
serial==0.0.97
six==1.16.0
sqlparse==0.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.4
tensorflow-estimator==2.3.0
termcolor==1.1.0
tornado==6.1
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
